Reorganization of header files. Minor fixes.
