+++ CMakeLists.txt
@@ -7,10 +7,10 @@ find_package(Threads)
 include_directories(${PROJECT_SOURCE_DIR}/include)
 add_definitions(-std=c++11)
 
-add_executable(test1
+add_executable(tests
     tests/tests.cpp
-    include/ndimage.h
-    include/narrowband.h
-    include/morphsnakes.h
+    include/morphsnakes/ndimage.h
+    include/morphsnakes/narrowband.h
+    include/morphsnakes/morphsnakes.h
 )
 target_link_libraries(test1 ${CMAKE_THREAD_LIBS_INIT})
