+++ include/morphsnakes/narrowband.h
